+++ resource/opentelemetry-resource-detector-azure/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Ignore vm detector if already in other rps
   ([#2456](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2456))
+- Implement functions resource detector
+  ([#2456](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2456))
 
 ## Version 0.1.4 (2024-04-05)